Implemented a simple solver.
When generating a masked board, the solver is used to determine whether
the board with the given mask can easily be solved by a human, and ups
the number of shown cells until the board is easy enough.

Other changes include moving the cell finders from RandomBoard to the
Board class, shortening the implementation of digitCharToInt and also
moving it to the Board class, adding a new static method
"intToDigitChar" in the Board class, and renaming the finders to match
Java conventions for public, static, final fields.